Developer

+++ lib/screens/children/blocs/children_management_bloc/children_management_bloc.dart
@@ -1,5 +1,6 @@
 import 'package:bloc/bloc.dart';
 import 'package:child_repository/child_repository.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:perfect_childcare/screens/children/blocs/children_management_bloc/children_management_event.dart';
 import 'package:perfect_childcare/screens/children/blocs/children_management_bloc/children_management_state.dart';
 import 'package:user_repository/user_repository.dart';
@@ -19,21 +20,18 @@ class ChildrenManagementBloc
   }
 
   Future<void> _onAddChild(
-      AddChildEvent event, Emitter<ChildrenManagementState> emit) async {
+    AddChildEvent event,
+    Emitter<ChildrenManagementState> emit,
+  ) async {
     emit(ChildrenManagementLoading());
     try {
+      if (FirebaseAuth.instance.currentUser == null) {
+        emit(const ChildrenManagementFailure('User not logged in'));
+        return;
+      }
       await childRepository.addChild(event.child);
       // Update user's children list
       await userRepository.connectChildToUser(event.userId, event.child.id);
-      final user = await userRepository.getUserById(event.userId);
-
-      if (user != null) {
-        final linkedPersonId = user.linkedPerson;
-        if (linkedPersonId.isNotEmpty) {
-          await userRepository.connectChildToUser(
-              linkedPersonId, event.child.id);
-        }
-      }
 
       emit(ChildrenManagementSuccess());
     } catch (e) {
@@ -45,6 +43,10 @@ class ChildrenManagementBloc
       RemoveChildEvent event, Emitter<ChildrenManagementState> emit) async {
     emit(ChildrenManagementLoading());
     try {
+      if (FirebaseAuth.instance.currentUser == null) {
+        emit(const ChildrenManagementFailure('User not logged in'));
+        return;
+      }
       await userRepository.disconnectChildFromUser(event.childId);
       await childRepository.removeChild(event.childId);
 
@@ -58,6 +60,10 @@ class ChildrenManagementBloc
       UpdateChildEvent event, Emitter<ChildrenManagementState> emit) async {
     emit(ChildrenManagementLoading());
     try {
+      if (FirebaseAuth.instance.currentUser == null) {
+        emit(const ChildrenManagementFailure('User not logged in'));
+        return;
+      }
       await childRepository.updateChild(event.child);
       emit(ChildrenManagementSuccess());
     } catch (e) {
